refactor: create api endpoint partitions
