+++ src/app_modules/api/v1/routes.rs
@@ -1,0 +1,8 @@
+use actix_web::web;
+
+use super::user_endpoints;
+
+// Grouped routes for users
+pub fn user_routes(cfg: &mut web::ServiceConfig) {
+    cfg.service(web::scope("/users").service(user_endpoints::get_user));
+}
